requirements.txt
seleniumbase
pytest
pytest-html
PyYAML
markupsafe